Refactor candidate jobs page into a component
